// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export type Billing = {
  __typename?: 'Billing';
  address: Scalars['String']['output'];
  district?: Maybe<Scalars['String']['output']>;
  email?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  phone: Scalars['String']['output'];
};

export type BillingInput = {
  address: Scalars['String']['input'];
  district?: InputMaybe<Scalars['String']['input']>;
  email?: InputMaybe<Scalars['String']['input']>;
  name: Scalars['String']['input'];
  phone: Scalars['String']['input'];
};

export type Brand = {
  __typename?: 'Brand';
  _id?: Maybe<Scalars['ID']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  logo?: Maybe<ServerFileEntity>;
  name: Scalars['String']['output'];
  orgUID: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type ColorSchema = {
  __typename?: 'ColorSchema';
  color: Scalars['String']['output'];
  description: Scalars['String']['output'];
};
// ...
export type Employee = {
  __typename?: 'Employee';
  _id?: Maybe<Scalars['ID']['output']>;
  employeeDetails: User;
  organizations?: Maybe<Array<EmployeeOrganizationSchema>>;
};
// ...
export type LineItem = {
  __typename?: 'LineItem';
  code?: Maybe<Scalars['String']['output']>;
  color?: Maybe<Scalars['String']['output']>;
  price: Scalars['Float']['output'];
  product: ProductDetails;
  productId: Product;
  quantity: Scalars['Int']['output'];
  size?: Maybe<Scalars['String']['output']>;
  subtotal: Scalars['Float']['output'];
};
// ...
export enum OrderStatus {
  Cancelled = 'CANCELLED',
  Confirmed = 'CONFIRMED',
  Delivered = 'DELIVERED',
  Pending = 'PENDING',
  Processing = 'PROCESSING',
  Refunded = 'REFUNDED',
  Shipped = 'SHIPPED'
}
// ...
export type Payment = {
  __typename?: 'Payment';
  amount?: Maybe<Scalars['Float']['output']>;
  method: PaymentMethod;
  paidAt?: Maybe<Scalars['DateTime']['output']>;
  status: PaymentStatus;
  transactionId?: Maybe<Scalars['String']['output']>;
};

export type PaymentInput = {
  amount?: InputMaybe<Scalars['Float']['input']>;
  method: PaymentMethod;
  paidAt?: InputMaybe<Scalars['DateTime']['input']>;
  status: PaymentStatus;
  transactionId?: InputMaybe<Scalars['String']['input']>;
};

export enum PaymentMethod {
  CashOnDelivery = 'CASH_ON_DELIVERY',
  OnlinePayment = 'ONLINE_PAYMENT'
}

export enum PaymentStatus {
  Due = 'DUE',
  Paid = 'PAID',
  Refunded = 'REFUNDED',
  Unpaid = 'UNPAID'
}

export enum Payment_Status {
  Cancelled = 'CANCELLED',
  Due = 'DUE',
  Paid = 'PAID',
  PartiallyPaid = 'PARTIALLY_PAID',
  Refunded = 'REFUNDED'
}

export type Product = {
  __typename?: 'Product';
  _id?: Maybe<Scalars['ID']['output']>;
  brand?: Maybe<Brand>;
  carouselImages?: Maybe<Array<ServerFileEntity>>;
  category: ProductCategory;
  code: Scalars['String']['output'];
  colors?: Maybe<Array<ColorSchema>>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  discountAmount?: Maybe<Scalars['Float']['output']>;
  gallery?: Maybe<Array<ServerFileEntity>>;
  model?: Maybe<Scalars['String']['output']>;
  orgUID: Scalars['String']['output'];
  regularPrice: Scalars['Float']['output'];
  salePrice?: Maybe<Scalars['Float']['output']>;
  shortDescription?: Maybe<Scalars['String']['output']>;
  sizes?: Maybe<Array<SizeSchema>>;
  stock: Scalars['Float']['output'];
  stockHistory?: Maybe<Array<StockHistorySchema>>;
  thumbnail?: Maybe<ServerFileEntity>;
  title: Scalars['String']['output'];
  unit?: Maybe<Unit>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};

export type ProductCategoriesListQueryDto = {
  limit?: InputMaybe<Scalars['Int']['input']>;
  page?: InputMaybe<Scalars['Int']['input']>;
  sort?: InputMaybe<SortType>;
  sortBy?: InputMaybe<Scalars['String']['input']>;
  where?: InputMaybe<Array<CommonMatchInput>>;
  whereOperator?: InputMaybe<Where_Operator>;
};

export type ProductCategory = {
  __typename?: 'ProductCategory';
  _id?: Maybe<Scalars['ID']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  orgUID: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};

export type ProductCategoryPagination = {
  __typename?: 'ProductCategoryPagination';
  meta?: Maybe<PaginationMeta>;
  nodes?: Maybe<Array<ProductCategory>>;
};

export type ProductDetails = {
  __typename?: 'ProductDetails';
  code: Scalars['String']['output'];
  model?: Maybe<Scalars['String']['output']>;
  orgUID: Scalars['String']['output'];
  regularPrice: Scalars['Float']['output'];
  salePrice?: Maybe<Scalars['Float']['output']>;
  thumbnail?: Maybe<ServerFileEntity>;
  title: Scalars['String']['output'];
};
// ...
export type ServerFileEntity = {
  __typename?: 'ServerFileEntity';
  bucket?: Maybe<Scalars['String']['output']>;
  externalUrl?: Maybe<Scalars['String']['output']>;
  key?: Maybe<Scalars['String']['output']>;
  region?: Maybe<Scalars['String']['output']>;
};
// ...
export type SizeSchema = {
  __typename?: 'SizeSchema';
  description: Scalars['String']['output'];
  size: Scalars['String']['output'];
};

export enum SortType {
  Asc = 'ASC',
  Desc = 'DESC'
}

export type StockHistoryInput = {
  date: Scalars['DateTime']['input'];
  quantity: Scalars['Int']['input'];
  stockPrice: Scalars['Float']['input'];
  stockType: StockType;
  updatedBy: Scalars['String']['input'];
};

export type StockHistorySchema = {
  __typename?: 'StockHistorySchema';
  date: Scalars['DateTime']['output'];
  quantity: Scalars['Int']['output'];
  stockPrice: Scalars['Float']['output'];
  stockType: StockType;
  updatedBy: Employee;
};

export enum StockType {
  StockIn = 'Stock_In',
  StockOut = 'Stock_Out'
}
// ...
export type Unit = {
  __typename?: 'Unit';
  _id?: Maybe<Scalars['ID']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  name: Scalars['String']['output'];
  orgUID: Scalars['String']['output'];
  unitCode: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};